#30 feat : challengeDetail data mapper 구현
