--- app/src/main/java/com/aoztg/greengrim/presentation/ui/challenge/model/ChallengeDetail.kt
@@ -1,10 +1,0 @@
-package com.aoztg.greengrim.presentation.ui.challenge.model
-
-data class ChallengeDetail(
-    val imgUrl: String = "",
-    val title: String = "",
-    val description: String = "",
-    val mainKeyword: List<String> = emptyList(),
-    val subKeyword: List<String> = emptyList(),
-    val date: String = "",
-)
