+++ app/src/main/java/com/aoztg/greengrim/presentation/ui/global/mapper/ChallengeDetailMapper.kt
@@ -1,0 +1,16 @@
+package com.aoztg.greengrim.presentation.ui.global.mapper
+
+import com.aoztg.greengrim.data.model.ChallengeDetailResponse
+import com.aoztg.greengrim.presentation.ui.global.model.ChallengeDetail
+
+
+fun ChallengeDetailResponse.toChallengeDetail(): ChallengeDetail{
+    return ChallengeDetail(
+        id = this.challengeInfo.id,
+        imgUrl = this.challengeInfo.imgUrl,
+        title = this.challengeInfo.title,
+        description = this.challengeInfo.description,
+        keywords = this.challengeTags,
+        date = this.createdAt
+    )
+}
