+++ app/src/main/java/com/aoztg/greengrim/presentation/ui/global/model/ChallengeDetail.kt
@@ -1,0 +1,12 @@
+package com.aoztg.greengrim.presentation.ui.global.model
+
+import com.aoztg.greengrim.data.model.ChallengeDetailTags
+
+data class ChallengeDetail(
+    val id: Int,
+    val imgUrl: String = "",
+    val title: String = "",
+    val description: String = "",
+    val keywords: ChallengeDetailTags?=null,
+    val date: String = "",
+)
